backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
typing-extensions==4.9.0
